requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
coverage==5.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.8
django-throttle-requests==0.5.9
flake8==3.8.3
future==0.18.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
mccabe==0.6.1
names==0.3.0
psycopg2==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
redis==3.5.3
redis-server==5.0.7
requests==2.24.0
sqlparse==0.3.1
static3==0.7.0
urllib3==1.25.10
vine==1.3.0
zipp==3.1.0
